+++ flowman-core/src/test/scala/com/dimajix/flowman/types/IntegerTypeTest.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Kaya Kupferschmidt
+ * Copyright 2018-2022 Kaya Kupferschmidt
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -102,7 +102,7 @@ class IntegerTypeTest extends AnyFlatSpec with Matchers {
     }
 
     it should "provide the correct SQL type" in {
-        IntegerType.sqlType should be ("integer")
+        IntegerType.sqlType should be ("int")
         IntegerType.sparkType.sql should be ("INT")
         IntegerType.typeName should be ("integer")
     }